Moved to a "module" style directory structure, where the routes and controllers for any resource can be found together
