+++ src/components/post/post-routes.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import * as posts from '../controllers/post-controller';
+import * as posts from './post-controller';
 
 const router = express.Router(); // eslint-disable-line new-cap
