+++ src/components/user/user-routes.js
@@ -1,6 +1,6 @@
 import express from 'express';
-import * as users from '../controllers/user-controller';
-import { csbTokenAuthenticatedOnly } from '../security/authentication/csb-token-authentication';
+import * as users from './user-controller';
+import { csbTokenAuthenticatedOnly } from '../authentication/csb-token-authentication';
 
 const router = express.Router(); // eslint-disable-line new-cap
